Fix inconsistency of dec() function between builds and fail if bad input instead of returning random value
dec() was returning different outputs for the same string value on different platforms.
For example, output of dec("aaaaaaaaaaaaaaaa") is 12297829382473033728.00 on Linux-64bit but it is 2863311530.00 on Linux-32bit and Windows-64bit.
The issue was caused by the use of (unsigned long) type which has different size depending on the build platform.
The solution is to use (unsigned long long) which is guaranteed to be 64-bit on all platforms.

Also if the input to dec() is not a valid hexadecimal string, it was returning a random value because sscanf doesn't set x in this case and the code didn't check the value returned by sscanf.
The solution is to return an error if sscanf returns 0.