Speedup string append (10x to 800x) by using realloc instead of malloc/free and fix ring_state_realloc implementation
The code was using malloc-copy-free strategy which is not optimal since C runtime provides an optimized realloc implementation that gives huge speedup in scenarios where we need to append data to an existing pointer.
Moreover, the implementation of ring_state_realloc was not correct since it didn't handle the case of pointers belonging to the internal memory pool causing issues.

The example below demonstrates the performance gain: on Linux 64bit, before change it gives a 8.17 seconds performance and after the change it gives 0.01 seconds.

fullStr = "initial"
lineStr = "Dummy content line"

t1 = clock()
for i = 1 to 20000 {
	fullStr+= lineStr
}
t2 = clock()

delta = (t2 - t1) /clockspersecond()

? "total length = " + len(fullStr) + " characters"
? "perf= " + delta + " seconds"