Fix bug in isWindows64 returning 0 when ring.exe is a 64-bit binary
If ring.exe is 64-bit, then obviously Windows is 64-bit but the code was calling IsWow64Process which returns false if the process is 64-bit since in this case Wow64 is not enabled for this process (it is native 64-bit).

The fix is to use the define _WIN64 which exists when building in 64-bit mode (it is defined by both visual studio c++ and mingw64-g++).