Fix inconsistency of "unsigned" function between Linux and Windows builds
"Unsigned" function was not returning the same results between Linux and Windows. This was caused by the fact that it was using "usigned long" type which doesn't have the same size between Windows and Linux.

The below program display different result between Linux and Windows:

a = 4_294_967_295
b = 1
? unsigned (a,b, "+")

On Windows it display "0" while on Linux it displays  4294967296

The fix is easy: use RING_UNSIGNEDLONGLONG type instead of "unsigned long"