RingLibUI: Fix memory leaks and memory corruption related to wrong usage of uiFreeText
The sample test3.ring was crashing on Linux because of call to uiFreeText. This function should not be called from Ring code since it is an internal libUI function responsible for freeing memory returned by various functions (like uiOpenFile) and it should not be called on Ring string pointers.

Moreover, many libUI functions returning "char *" were leaking memory because uiFreeText was not called after copying the string to Ring.

This change addresses the memory leaks and removes uiFreeText from the list of libUI function visible to Ring code.