Enhancements to Windows build batch files: better usage of setlocal and use 'call' in locatevc.bat when running buildarch.ring
The setlocal change avoids that environment variables are polluted which was causing failure when building many extensions in the same CMD session.
The locatevc.bat change fixes an issue when using "auto" as first parameter because running "RINGEXEPATH" at line 32 was causing an error in some cases: using call avoid this problem and it is more reliable.