Update StdLib - Add AppArguments() function
AppArguments function returns the effective arguments passed to the Ring script. This is more practical than sysargv when we just need the effective arguments specified by the user to the Ring script and not the exe name and script name. Moreover, depending if the Ring script is compiled or not, sysargv doesn't have the same format with regards to the position of the effective arguments.