RingOpenSSL: enhancements to support all cipher algorithms and remove encrypt_v2.c limitation to 1.1 version
For of all, the code in encrypt_v2.c was working for openssl version 0.9.8 and up, so there was no need to limit it to version 1.1. encrypt_v1.c is kept for very old OpenSSL version (0.9.7 and older).
Encrypt and Decrypt function now accept a 4th parameter indicating the cipher algorithm to use. If not specified, then Blowfish in CBC mode is used for compatibility with older Ring versions
The code in encrypt_v2.c is enhanced in order to use OpenSSL functions in a better way that allows using different type of ciphers seamlessly. Also internal memory allocation is fixed because it was wrongly set to always allocate double the size of input which was too much or too little depending on cases.
A new function supportedciphers() is added and which returns the list of cipher algorithms supported for Encrypt/Decrypt functions.
A test script test_encrypt_decrypt.ring is added: it runs test vectors for popular ciphers (AES, DES, TripleDES and Blowfish) in various modes.
stdlib library is enhanced with functions related to these new additions.
The encrypt/decrypt sample is the documentation is modernized to use AES-128.