RingOpenSSL: first implementation of RSA support (key generation,import,export,encryption,decryption)
The script test_rsa.ring implements tests for all added functions and it can serve as a reference alongside the documentation.
Below is the list of added functions. The documentation (secfunc.txt) was updated with all details about these functions and examples of usage.
 - rsa_generate
 - rsa_export_params
 - rsa_import_params
 - rsa_export_pem
 - rsa_import_pem
 - rsa_is_privatekey
 - rsa_encrypt_pkcs
 - rsa_decrypt_pkcs
 - rsa_encrypt_oaep
 - rsa_decrypt_oaep
 - openssl_versiontext
 - openssl_version

One interesting point about test_rsa.ring is that I implemented a deep comparison function for lists (cf Compare in test_rsa.ring). It would have been helpful to have such function in the standard library.