Fix failure of SystemCmd function if current directory is not writable
SystemCmd implementation redirected command output to cmd.txt file in current directory which caused an error if the current directory is not writable by the current process.
Now we use the tempname() function to create a dedicated random temporary file for capturing command output which is guaranteed to be always writable.