Tests: fix wrong expected output of general_fread test by removing erroneous new lines
The expected output of general\fread.ring test was wrong because it had erroneous new lines added to it.
This made the test appear as failing although it was succeeding.
This commit remove the erroneous new lines so that the output matches the content of original fread.ring file as stored in git repository.